Rally base garrison to defend

+++ code/src/My/MainBundle/Controller/CommandsController.php
@@ -62,6 +62,9 @@ class CommandsController extends Controller
 
         $em = $this->getDoctrine()->getManager();
         foreach ($fleets as $f) {
+            if ($f->getPlayer != $player)
+                return $this->fail("Leave the fleets alone");
+            
             if (in_array($f->getId(), $selectedFleets)) {
                 $base->addPower($f->getPower());
                 $em->remove($f);
